requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
google-api-core==1.11.0
google-auth==1.6.3
google-cloud-language==1.2.0
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.8
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.2
